+++ NQueen.Benchmarking/Usings.cs
@@ -9,5 +9,3 @@ global using NQueen.Domain.Models;
 global using NQueen.Domain.Utils;
 global using NQueen.Kernel.Solvers;
 global using NQueen.Kernel.Solvers.Engines;
-global using System;
-global using System.Linq;